pyproject.toml
[project]
name = "esgf-wut"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "typer>=0.19.2",
]

[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=42", "setuptools-scm>=7"]

[tool.setuptools]
packages = ["esgf_wut"]

[tool.setuptools_scm]
version_scheme = "no-guess-dev"
local_scheme = "node-and-date"
fallback_version = "0.0.0"
write_to = "esgf_wut/_version.py"

[tool.setuptools.package-data]
esgf_magic = ["data/*"]

[project.scripts]
wut = "esgf_wut.cli:app"
